refactor: split `Funnel Components` Declarations from `useFunnel`
