--- hooks/use-funnel/context.ts
@@ -1,7 +1,0 @@
-import { Dispatch, createContext } from 'react';
-
-export type FunnelContextProps = {
-  currentStep: string;
-};
-
-export const FunnelContext = createContext<FunnelContextProps | null>(null);
