+++ hooks/use-funnel/index.tsx
@@ -1,14 +1,8 @@
-import { ReactNode, useState } from 'react';
-import { FunnelContext } from './context';
-import { useFunnelContext } from './Funnel';
+import { ReactNode, useMemo, useState } from 'react';
 import Steps from './Step';
-import { AppRouterInstance } from 'next/dist/shared/lib/app-router-context';
+import Funnel from './Funnel';
 
-type FunnelProps = {
-  children: React.ReactNode;
-};
-
-export type ArrToUnion<T> = T extends ReadonlyArray<infer ITEM> ? ITEM : never;
+type ArrToUnion<T> = T extends ReadonlyArray<infer ITEM> ? ITEM : never;
 
 export function useFunnel<T extends ReadonlyArray<string>>(steps: T) {
   if (steps === null || steps.length === 0 || typeof steps === 'undefined') {
@@ -20,15 +14,14 @@ export function useFunnel<T extends ReadonlyArray<string>>(steps: T) {
     setStep(next);
   };
 
-  function Funnel({ children }: FunnelProps) {
-    return (
-      <FunnelContext.Provider value={{ currentStep }}>
-        {children}
-      </FunnelContext.Provider>
+  const FunnelComponent = useMemo(() => {
+    return Object.assign(
+      ({ children }: { children: ReactNode }) => {
+        return <Funnel currentStep={currentStep}>{children}</Funnel>;
+      },
+      { Steps: Steps<ArrToUnion<T>> }
     );
-  }
+  }, [currentStep]);
 
-  Funnel.Steps = Steps<ArrToUnion<T>>;
-
-  return [Funnel, currentStep, handleStep] as const;
+  return [FunnelComponent, currentStep, handleStep] as const;
 }
